+++ tfjs-core/src/ops/ops.ts
@@ -20,6 +20,7 @@ export {add} from './add';
 export {addN} from './add_n';
 export {all} from './all';
 export {any} from './any';
+export {argMax} from './arg_max';
 export {atan2} from './atan2';
 export {avgPool} from './avg_pool';
 export {avgPool3d} from './avg_pool_3d';
