+++ tfjs-core/src/public/chained_ops/register_all_chained_ops_test.ts
@@ -27,6 +27,7 @@ const CHAINED_OPS = [
   'add',
   'all',
   'any',
+  'argMax',
   'atan2',
   'avgPool',
   'batchNorm',
